fix(badge): add flex styles to badge
adding flex styles so it has better bahavior when used alongside other elements. the remove of py is
accounted by increasing line-height

BREAKING CHANGE: it now behaves as inline-flex instead of default inline span. py-1 was also removed
in favor of leading increase, so it better aligns in the center of the pill.