feat(dropdown): add align prop to dropdown
BREAKING CHANGE: Dropdown now accepts an align prop that can be 'right' or 'left'. It now defaults
to 'left' if align is not used, which is the natural align of the DOM, instead of the old right
alignment. Change your dropdowns to align="right" if you were using the old default style.