fix(form): fix 1px off input and select
regular inputs had a height of 39px, which would make it terrible when used with a regular button
(height 38px). reducing the line-height solved the problem.

BREAKING CHANGE: Regular Inputs and Selects had a height of 39px, which looked terrible when used
side-by-side with a regular button (38px height). Reducing the `line-height` to `leading-5` solved
it, and now regular inputs and selects have a 38px height.